Split quirk translator fork into modules
Remove the input, punctuation, and regex filter sections into their own modules.
Add in export and import code to connect
Add array declarations in each translator to fix new reference error of undeclared array.
